Prueba de peticiones

+++ src/main/java/com/uptc/frw/jpa/repository/OpcionesModeloRepository.java
@@ -3,6 +3,6 @@ package com.uptc.frw.jpa.repository;
 import com.uptc.frw.jpa.model.OpcionesModelo;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface OpcionesModeloRepository extends JpaRepository<OpcionesModelo, Long> {
+public interface OpcionesModeloRepository extends JpaRepository<OpcionesModelo, Integer> {
 }
